backend/requirements.txt
flask
flask_cors
boto3
python-dotenv
replicate
gunicorn
setuptools
wheel
playsound==1.2.2
